requirements.txt
pandas
numpy
scikit-learn
matplotlib
joblib
fastapi
uvicorn
keras
langchain
tqdm
python-dotenv
requests
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
unstructured
python-multipart
langchain_huggingface
seaborn